module verify-3rd-provider-api
change all files
