+++ third-party-verify/.gitattributes
