+++ third-party-verify/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'third-party-verify'
