+++ third-party-verify/src/main/java/vn/chuot96/verify3rd/ThirdPartyVerify.java
@@ -1,0 +1,12 @@
+package vn.chuot96.verify3rd;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ThirdPartyVerify {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ThirdPartyVerify.class, args);
+    }
+}
