+++ third-party-verify/src/main/java/vn/chuot96/verify3rd/compo/OauthParty.java
@@ -1,0 +1,9 @@
+package vn.chuot96.verify3rd.compo;
+
+import vn.chuot96.verify3rd.dto.UserDTO;
+
+public interface OauthParty {
+    String getProviderName();
+
+    UserDTO verifyToken(String token);
+}
