+++ third-party-verify/src/main/java/vn/chuot96/verify3rd/dto/UserDTO.java
@@ -1,0 +1,3 @@
+package vn.chuot96.verify3rd.dto;
+
+public record UserDTO(String provider, String subject) {}
