+++ third-party-verify/src/main/java/vn/chuot96/verify3rd/exception/InvalidTokenException.java
@@ -1,0 +1,11 @@
+package vn.chuot96.verify3rd.exception;
+
+public class InvalidTokenException extends RuntimeException {
+    public InvalidTokenException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public InvalidTokenException(String message) {
+        super(message);
+    }
+}
