+++ third-party-verify/src/main/resources/application-verify1.properties
@@ -1,0 +1,11 @@
+spring.application.name = third-party-verify-tbirpg2d
+#__________________________________________________
+server.port = 8001
+#__________________________________________________
+# ========== Eureka Server Registration ==========
+eureka.client.service-url.defaultZone = http://localhost:8761/eureka
+eureka.instance.prefer-ip-address = true
+eureka.instance.instance-id = ${spring.application.name}:${server.port}
+#__________________________________________________
+# ============== External File Path ===============
+external-file.path = E:/Document/texts/data.txt
