+++ third-party-verify/src/main/resources/application.properties
@@ -1,0 +1,3 @@
+spring.application.name = third-party-verify
+
+
