+++ third-party-verify/src/test/java/vn/chuot96/verify3rd/ThirdPartyVerifyTests.java
@@ -1,0 +1,11 @@
+package vn.chuot96.verify3rd;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ThirdPartyVerifyTests {
+
+    @Test
+    void contextLoads() {}
+}
