--- verify-3rd-provider-api/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'verify-3rd-provider-api'
