--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/Verify3rdProviderAPI.java
@@ -1,12 +1,0 @@
-package vn.chuot96.verify3rdapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Verify3rdProviderAPI {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Verify3rdProviderAPI.class, args);
-    }
-}
