--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/component/GoogleOauthProvider.java
@@ -1,20 +1,0 @@
-package vn.chuot96.verify3rdapi.component;
-
-import static vn.chuot96.verify3rdapi.constant.Provider.GOOGLE;
-
-import org.springframework.stereotype.Component;
-import vn.chuot96.verify3rdapi.dto.UserDTO;
-import vn.chuot96.verify3rdapi.util.OauthHandler;
-
-@Component("google")
-public class GoogleOauthProvider implements OauthProvider {
-    @Override
-    public String getProviderName() {
-        return GOOGLE.getKey();
-    }
-
-    @Override
-    public UserDTO verifyToken(String token) {
-        return OauthHandler.googleProvider(token);
-    }
-}
