--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/component/OauthProvider.java
@@ -1,9 +1,0 @@
-package vn.chuot96.verify3rdapi.component;
-
-import vn.chuot96.verify3rdapi.dto.UserDTO;
-
-public interface OauthProvider {
-    String getProviderName();
-
-    UserDTO verifyToken(String token);
-}
