--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/config/WebClientConfig.java
@@ -1,15 +1,0 @@
-package vn.chuot96.verify3rdapi.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-    @Bean
-    public WebClient.Builder webClientBuilder() {
-        return WebClient.builder().defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE);
-    }
-}
