--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/constant/Provider.java
@@ -1,19 +1,0 @@
-package vn.chuot96.verify3rdapi.constant;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Provider {
-    GOOGLE("google", "https://www.googleapis.com/oauth2/v3/certs"),
-    FACEBOOK("facebook", "https://graph.facebook.com/me?fields=id,name,email,picture&access_token="),
-    FIREBASE("eyJ", "https://securetoken.google.com/PROJECT_ID");
-
-    //    AUTH0("auth0", "https://AUTH0_DOMAIN/"),
-    //    COGNITO("cognito", "https://cognito-idp.REGION.amazonaws.com/USER_POOL_ID"),
-    //    AZURE("azure", "https://login.microsoftonline.com/TENANT_ID/v2.0");
-
-    private final String key;
-    private final String endpoint;
-}
