--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/dto/RequestDTO.java
@@ -1,4 +1,0 @@
-package vn.chuot96.verify3rdapi.dto;
-
-public record RequestDTO(String appId, String token) {
-}
