--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/dto/UserDTO.java
@@ -1,3 +1,0 @@
-package vn.chuot96.verify3rdapi.dto;
-
-public record UserDTO(String provider, String userId) {}
