--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/exception/InvalidTokenException.java
@@ -1,11 +1,0 @@
-package vn.chuot96.verify3rdapi.exception;
-
-public class InvalidTokenException extends RuntimeException {
-    public InvalidTokenException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public InvalidTokenException(String message) {
-        super(message);
-    }
-}
