--- verify-3rd-provider-api/src/main/java/vn/chuot96/verify3rdapi/exception/NoSupportedProviderException.java
@@ -1,11 +1,0 @@
-package vn.chuot96.verify3rdapi.exception;
-
-public class NoSupportedProviderException extends RuntimeException {
-    public NoSupportedProviderException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public NoSupportedProviderException(String message) {
-        super(message);
-    }
-}
