--- verify-3rd-provider-api/src/main/resources/application.properties
@@ -1,13 +1,0 @@
-spring.application.name = verify-3rd-provider-api
-#________________________________________________
-server.port = 8201
-#________________________________________________
-# eureka server registration:
-eureka.client.service-url.defaultZone = http://localhost:8761/eureka
-eureka.instance.prefer-ip-address = true
-eureka.instance.instance-id = ${spring.application.name}:${server.port}
-#________________________________________________
-# external file path:
-external-file.path = E:/Document/texts/verify-3rd-provider-api.txt
-
-
