--- verify-3rd-provider-api/src/test/java/vn/chuot96/authen3rdAPI/Authen3RdAPITests.java
@@ -1,11 +1,0 @@
-package vn.chuot96.authen3rdAPI;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Authen3RdAPITests {
-
-    @Test
-    void contextLoads() {}
-}
